feat: refactor duplicated code of otlp exporter
