+++ exporters/otlp/internal/go.mod
@@ -1,20 +1,16 @@
-module go.opentelemetry.io/otel
+module go.opentelemetry.io/otel/exporters/otlp/internal
 
 go 1.18
 
 require (
-	github.com/go-logr/logr v1.2.3
-	github.com/go-logr/stdr v1.2.2
-	github.com/google/go-cmp v0.5.8
-	github.com/stretchr/testify v1.7.1
+	github.com/stretchr/testify v1.8.0
 	go.opentelemetry.io/otel/exporters/otlp/internal/retry v1.10.0
-	go.opentelemetry.io/otel/trace v1.10.0
 	google.golang.org/grpc v1.49.0
 )
 
 require (
 	github.com/cenkalti/backoff/v4 v4.1.3 // indirect
-	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
@@ -22,7 +18,5 @@ require (
 	golang.org/x/text v0.3.3 // indirect
 	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
 	google.golang.org/protobuf v1.27.1 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace go.opentelemetry.io/otel/trace => ./trace
