--- exporters/otlp/otlpmetric/internal/oconf/optiontypes.go
@@ -1,55 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package oconf // import "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal/oconf"
-
-import "time"
-
-const (
-	// DefaultCollectorGRPCPort is the default gRPC port of the collector.
-	DefaultCollectorGRPCPort uint16 = 4317
-	// DefaultCollectorHTTPPort is the default HTTP port of the collector.
-	DefaultCollectorHTTPPort uint16 = 4318
-	// DefaultCollectorHost is the host address the Exporter will attempt
-	// connect to if no collector address is provided.
-	DefaultCollectorHost string = "localhost"
-)
-
-// Compression describes the compression used for payloads sent to the
-// collector.
-type Compression int
-
-const (
-	// NoCompression tells the driver to send payloads without
-	// compression.
-	NoCompression Compression = iota
-	// GzipCompression tells the driver to send payloads after
-	// compressing them with gzip.
-	GzipCompression
-)
-
-// RetrySettings defines configuration for retrying batches in case of export failure
-// using an exponential backoff.
-type RetrySettings struct {
-	// Enabled indicates whether to not retry sending batches in case of export failure.
-	Enabled bool
-	// InitialInterval the time to wait after the first failure before retrying.
-	InitialInterval time.Duration
-	// MaxInterval is the upper bound on backoff interval. Once this value is reached the delay between
-	// consecutive retries will always be `MaxInterval`.
-	MaxInterval time.Duration
-	// MaxElapsedTime is the maximum amount of time (including retries) spent trying to send a request/batch.
-	// Once this value is reached, the data is discarded.
-	MaxElapsedTime time.Duration
-}
